fixes ref #59 
Adds support for SoAs of arbitrary types:
All that needs to be done is to add two things to your particle class:

```C++
enum AttributeNames : size_t { id, posX, posY, posZ, forceX, forceY, forceZ };
typedef autopas::utils::SoAType<size_t, double, double, double, double, double, double>::Type SoAArraysType;
```

The `enum AttributeNames` specifies which variables are allowed. The `typedef SoAArraysType` specifies the type of the SoAs, please add the corresponding types of the fields in attributenames.
Both have to have the same number of arguments.

+++ src/iterators/SingleCellIterator.h
@@ -92,7 +92,7 @@ class SingleCellIterator : public SingleCellIteratorInterfaceImpl<Particle> {
    * Get the index of the particle in the cell
    * @return index of the current particle
    */
-  int getIndex() const override { return _index; }
+  size_t getIndex() const override { return _index; }
 
   /**
    * Deletes the current particle
